trabajo 24/10

+++ PSP/1erTRIMESTRE/worksapce-multihilo/src/ejerciciosT3/Caracol3.java
@@ -6,7 +6,6 @@ public class Caracol3 implements Runnable {
     private String nomCaracol;
     private double velCaracol;
     private boolean meta = false;
-    private static Thread[] hilos;
 
     public Caracol3(String n, double v) {
         nomCaracol = n; velCaracol = v;
@@ -16,32 +15,49 @@ public class Caracol3 implements Runnable {
         File fichero = new File(f);
         int i = 0;
         while (i < 100) {
-            i+=velCaracol;
-            if (i > 100) i=100; //si i se pasa de 100 en algun momento lo asignamos d vuelta a 100.
-            IO.println("El caracol "+nomCaracol+" lleva un "+(int) Math.floor(i)+"% del total recorrido");
-            if (i == 100 && !meta) {
-                try {
-                    BufferedWriter bw = new BufferedWriter(new FileWriter(fichero));
-                    bw.write("El caracol "+nomCaracol+" ha llegado a la meta");
-                    bw.close();
-                    meta=true;
+            if (fichero.exists()) {
+                IO.println("El caracol "+nomCaracol + " abandona la carrera en el " +(int) Math.floor(i) +"%");
+                break;
+            } else {
+                i+=velCaracol;
+                if (i > 100) i=100; //si i se pasa de 100 en algun momento lo asignamos d vuelta a 100.
+                IO.println("El caracol "+nomCaracol+" lleva un "+(int) Math.floor(i)+"% del total recorrido");
+                if (i == 100 && !meta) {
+                    try {
+                        //Si no existe, bufferedWriter crea el fichero.
+                        BufferedWriter bw = new BufferedWriter(new FileWriter(fichero));
+                        IO.println("El caracol "+nomCaracol+" ha llegado a la meta");
+                        bw.write("El caracol "+nomCaracol+" ha llegado a la meta");
+                        bw.close();
+                        meta=true;
 
-                    for (Thread t: hilos) {
-                        if (t != Thread.currentThread()) {
-                            t.interrupt();
-                        }
+                    } catch (Exception e) {
+                        e.printStackTrace();
                     }
-
-                } catch (Exception e) {
+                }
+                try {
+                    Thread.sleep(50);
+                } catch (InterruptedException e) {
                     e.printStackTrace();
                 }
             }
-            try {
-                Thread.sleep(50);
-            } catch (InterruptedException e) {
-                e.printStackTrace();
-            }
         }
+       // String contenidoFicheroGanador = LeerFichero(fichero);
+        // IO.println(contenidoFicheroGanador);
+    }
+
+    public String LeerFichero(File f) {
+        String contenido="";
+        try (BufferedReader br=new BufferedReader(new FileReader(f))) {
+            String linea;
+            while ((linea = br.readLine()) != null) {
+                contenido+=linea+"\n";
+            }
+            br.close();
+        } catch (Exception e) {
+            e.printStackTrace();
+        }
+        return contenido;
     }
 
 
@@ -53,10 +69,8 @@ public class Caracol3 implements Runnable {
         Thread c1 = new Thread(new Caracol3("Pedro", 1+Math.random()*4));
         Thread c2 = new Thread(new Caracol3("Paco", 1+Math.random()*4));
         Thread c3 = new Thread(new Caracol3("Tate", 1+Math.random()*4));
-        Thread c4 = new Thread(new Caracol3("Nigga", 1+Math.random()*4));
-        Thread c5 = new Thread(new Caracol3("Porro", 1+Math.random()*4));
-
-        hilos = new Thread[] { c1, c2, c3, c4, c5 };
+        Thread c4 = new Thread(new Caracol3("Jeison", 1+Math.random()*4));
+        Thread c5 = new Thread(new Caracol3("Tilin", 1+Math.random()*4));
 
         c1.start();
         c2.start();
